requirements.txt
langgraph
langchain-core
pydantic
python-dotenv
anthropic
streamlit
pymongo[srv]
PyMuPDF
